+++ molecule/driver/ec2.py
@@ -92,7 +92,7 @@ class EC2(base.Base):
         driver:
           name: ec2
           ssh_connection_options:
-            -o ControlPath=~/.ansible/cp/%r@%h-%p
+            - '-o ControlPath=~/.ansible/cp/%r@%h-%p'
 
     .. important::