+++ CHANGELOG.rst
@@ -18,6 +18,7 @@ Unreleased
 * Add SSH password to delegated driver instance_dict
 * Add WinRM connections options to delegated driver instance_dict
 * Update testinfra to 3.0.6 so we can use ansible verbosity
+* Add ``sysctls`` option to the Docker driver.
 
 2.20
 ====